Add more tests and verification of input parameters verification

+++ common/include/FloatingPointUtils.hpp
@@ -6,7 +6,7 @@
 
 namespace spacey::common {
 
-inline auto areFloatingPointEquals(const double& first, const double& second) -> bool {
+inline auto areFloatingPointEquals(const long double& first, const long double& second) -> bool {
     return std::fabs(first - second) < std::numeric_limits<double>::epsilon();
 }
